Genericize simple stacked bar
